Last commit missed a couple files, here they are
